requirements.txt
Django==4.2.7
graphene-django==3.1.5
whitenoise>=6.0.0
django-graphql-jwt==0.3.4
django-cors-headers==4.2.0
django-filter==23.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.0.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
